+++ src/funcTest/groovy/info/solidsoft/gradle/pitest/functional/Junit5FunctionalSpec.groovy
@@ -36,6 +36,7 @@ class Junit5FunctionalSpec extends AbstractPitestFunctionalSpec {
 
     @Issue(["https://github.com/szpak/gradle-pitest-plugin/issues/177", "https://github.com/szpak/gradle-pitest-plugin/issues/300",
         "https://github.com/szpak/gradle-pitest-plugin/issues/337"])
+//    @IgnoreIf({ !jvm.isJavaVersionCompatible(21) && expectedJUnitJupiterVersion == "5.8.0" }) //TODO: Explain why it fails with "No such property: expectedJUnitJupiterVersion for class"
     void "should work with junit5 without explicitly adding dependency (#description)"() {
         given:
             copyResources("testProjects/junit5simple", "")
@@ -53,9 +54,10 @@ class Junit5FunctionalSpec extends AbstractPitestFunctionalSpec {
             result.standardOutput.contains("junit-jupiter-api-${expectedJUnitJupiterVersion}.jar")
             result.standardOutput.contains("junit-platform-commons-${expectedJUnitPlatformVersion}.jar")
         where:
-            buildFileName                             || expectedJunitPluginVersion | expectedJUnitJupiterVersion | expectedJUnitPlatformVersion
-            'build.gradle'                            || "1.2.0"                    | "5.10.0"                    | "1.10.0"
-            'build-pit-plugin-1.0.0-junit-5.8.gradle' || "1.0.0"                    | "5.8.0"                     | "1.8.0"
+            buildFileName                              || expectedJunitPluginVersion | expectedJUnitJupiterVersion | expectedJUnitPlatformVersion
+            'build.gradle'                             || "1.2.3"                    | "5.12.2"                    | "1.12.2"
+            'build-pit-plugin-1.2.0-junit-5.10.gradle' || "1.2.0"                    | "5.10.0"                    | "1.10.0"
+//            'build-pit-plugin-1.0.0-junit-5.8.gradle'  || "1.0.0"                    | "5.8.0"                     | "1.8.0"
 
             description = "plugin $expectedJunitPluginVersion, junit $expectedJUnitJupiterVersion, platform $expectedJUnitPlatformVersion"
     }